Release v1.3.6.6 (#4304)
+ Logs are now stored inside a profile folder \Logs\<profile_name>\*.txt
+ There's a General Log inside \Logs\GeneralLog.txt. Please note that other files are still stored in Configs dir (pokemons.txt, HuntStats...)
= Reduce log check disk requests
+ PTC Login process totally reviewed
+ ReAuthenticate process reviewed
+ Bot now use Cookies for authentication
+ Added new HEADER values on Authentication
+ In player tab you can see account creation date
+ Added a commandline argument to bypass the KillSwitch (at your own risk)
+ Added a commandline argument to select which profile to run (-nogui:<profile>)
= Fixed an exception when fort file to load does not exists
= Fixed some exception handles
= Minor fixes on Advance Breaks
= Some text fixes from spangish (Thanks Zederia)
= Some internal cleaning too

+++ PokemonGo.RocketAPI.Console/Program.cs
@@ -32,10 +32,7 @@ namespace PokeMaster
         public static string deviceData = Path.Combine(path_device, "DeviceData.json");
         public static string cmdCoords = string.Empty;
         public static string accountProfiles = Path.Combine(path, "Profiles.txt");
-        static readonly string logPath = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Logs");
-        public static string pokelog = Path.Combine(logPath, "PokeLog.txt");
-        public static string manualTransferLog = Path.Combine(logPath, "TransferLog.txt");
-        public static string EvolveLog = Path.Combine(logPath, "EvolveLog.txt");
+        public static string logPath = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Logs");
         public static string path_pokedata = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "PokeData");
 
         static Version version;       
@@ -116,6 +113,7 @@ namespace PokeMaster
                             }
                         }
 
+                        Logger.ColoredConsoleWrite(ConsoleColor.Red, "Using Profile: " + GlobalVars.ProfileName);
                         if (GlobalVars.UsePwdEncryption) GlobalVars.Password = Encryption.Decrypt(GlobalVars.Password);
                         #endregion
                     }
@@ -185,9 +183,9 @@ namespace PokeMaster
             }
 
             // Check if a new version of BOT is available
-            CheckVersion(); 
+            CheckVersion();
 
-            // Check if Bot is deactivated at server level
+             // Check if Bot is deactivated at server level
             if (!GlobalVars.BypassKillSwitch) StringUtils.CheckKillSwitch();
 
             if (openGUI)
@@ -212,7 +210,9 @@ namespace PokeMaster
             }
             
             SleepHelper.PreventSleep();
-            CreateLogDirectories();
+
+            // Ensure all log paths exists
+            CheckLogDirectories(logPath = Path.Combine(logPath, GlobalVars.ProfileName));
 
             GlobalVars.infoObservable.HandleNewHuntStats += SaveHuntStats;
 
@@ -249,7 +249,8 @@ namespace PokeMaster
             }
             else
             {
-                   System.Console.ReadLine();
+                Console.WriteLine("Press any key to continue.");
+                Console.ReadKey();
             }
             SleepHelper.AllowSleep();
         }
@@ -312,24 +313,20 @@ namespace PokeMaster
                         "https://raw.githubusercontent.com/Logxn/PokemonGo-Bot/master/ver.md");
         }
 
-        private static void CreateLogDirectories()
+        private static void CheckLogDirectories(string logPath)
         {
-            if (!Directory.Exists(logPath))
-            {
-                Directory.CreateDirectory(logPath);
-            }
-            if (!File.Exists(pokelog))
-            {
-                File.Create(pokelog).Close();
-            }
-            if (!File.Exists(manualTransferLog))
-            {
-                File.Create(manualTransferLog).Close();
-            }
-            if (!File.Exists(EvolveLog))
-            {
-                File.Create(EvolveLog).Close();
-            }
+            Logger.SwitchToProfileLog(logPath); // Here we change from General Log to profile log, stored inside Logs\<profile_name>\
+
+            string pokelog = Path.Combine(logPath, "PokeLog.txt");
+            string manualTransferLog = Path.Combine(logPath, "TransferLog.txt");
+            string EvolveLog = Path.Combine(logPath, "EvolveLog.txt");
+
+            if (!Directory.Exists(logPath)) Directory.CreateDirectory(logPath);
+            if (!File.Exists(pokelog)) File.Create(pokelog).Close();
+            if (!File.Exists(manualTransferLog)) File.Create(manualTransferLog).Close();
+            if (!File.Exists(EvolveLog)) File.Create(EvolveLog).Close();
+
+            Logger.Rename(logPath); // Rotate logs if checked
         }
     }
 }
